Add testing setup with Jest
This commit sets up a few preliminary unit tests and modularizes the
existing code just enough for the tests to work. There is much more to
do here!

The README now contains instructions for installing the node modules and
running the tests, but in short, once this commit is applied, you must
first install the required node modules, like this:

```
npm install
```

*Testing:*

You can run the tests like this:

```
npm run test
```

After the tests run, you will see a coverage report printed to the
screen and generated in the coverage/ folder. You can read more about
how to use the jest testing framework at https://facebook.github.io/jest/.

*Using the code in the browser:*

This commit adds module exports to the existing engine components. This
means that the components are now more easily testable, but it also
means that they must be specially packaged before they will run in the
browser.

*To use this library in the browser,* first run the build task, which
uses the browserify tool to bundle all the modules together into a
single browser-ready file:

```
npm run build
```

This generates a `build.js` file. This is the file you will include in
your HTML source. Once you've done that, all of the engine's components
are accesible on the `window.Engine` object in the browser (for example,
`Vector` is now `Engine.Vector`). The *test.html* file has been updated
to demonstrate this behavior.